feat(kit-ui): move styles folder in src
