+++ apps/front-client/src/pages/_app.tsx
@@ -3,11 +3,13 @@ import { AppProps } from 'next/app';
 import Head from 'next/head';
 import { ReactElement, ReactNode } from 'react';
 import { NextPage } from 'next';
+import { ThemeProvider } from '@emotion/react';
 import {
   Hydrate,
   QueryClient,
   QueryClientProvider,
 } from '@tanstack/react-query';
+import { muiTheme } from '@hetic-saline-royale-academy/kit-ui';
 
 export type NextPageWithLayout = NextPage & {
   getLayout?: (page: ReactElement) => ReactNode;
@@ -29,7 +31,11 @@ function ClientApp({ Component, pageProps }: AppPropsWithLayout) {
       </Head>
       <QueryClientProvider client={queryClient}>
         <Hydrate state={pageProps.dehydratedState}>
-          <main className="app">{getLayout(<Component {...pageProps} />)}</main>
+          <ThemeProvider theme={muiTheme}>
+            <main className="app">
+              {getLayout(<Component {...pageProps} />)}
+            </main>
+          </ThemeProvider>
         </Hydrate>
       </QueryClientProvider>
     </>
