+++ libs/kit-ui/.storybook/preview.tsx
@@ -1,7 +1,8 @@
+import React from 'react';
 import { withThemeFromJSXProvider } from '@storybook/addon-styling';
-import { GlobalStyles } from '../styles/global-style';
+import { GlobalStyles } from '../src/lib/styles/global-style';
 import { ThemeProvider } from '@emotion/react';
-import { muiTheme } from '../styles/theme';
+import { muiTheme } from '../src/lib/styles/theme';
 
 export const withMUITheme = (Story) => {
   return (
