+++ libs/kit-ui/src/lib/components/Atoms/Text/index.ts
@@ -1,3 +1,3 @@
 import Text from './Text';
 
-export { Text };
+export default Text;
