+++ libs/kit-ui/src/lib/index.ts
@@ -1,4 +1,5 @@
 export * from './components';
+export * from './styles/theme';
 
 export * from './palette';
 export * from './typography';
