+++ libs/kit-ui/src/lib/styles/mixins/apply-percent-ratio.ts
